requirements.txt
twelvelabs
snowflake-connector-python
streamlit


 subprocess
 youtube-transcript-api 
 google-generativeai 
